--- core/src/main/java/com/netflix/conductor/core/event/WorkflowCreationEvent.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2022 Conductor Authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package com.netflix.conductor.core.event;
-
-import java.io.Serializable;
-
-import com.netflix.conductor.core.execution.StartWorkflowInput;
-
-public class WorkflowCreationEvent implements Serializable {
-
-    private final StartWorkflowInput startWorkflowInput;
-
-    public WorkflowCreationEvent(StartWorkflowInput startWorkflowInput) {
-        this.startWorkflowInput = startWorkflowInput;
-    }
-
-    public StartWorkflowInput getStartWorkflowInput() {
-        return startWorkflowInput;
-    }
-}
