+++ core/src/main/java/com/netflix/conductor/core/execution/WorkflowExecutor.java
@@ -159,4 +159,10 @@ public interface WorkflowExecutor {
      * @param workflow workflow
      */
     void scheduleNextIteration(TaskModel task, WorkflowModel workflow);
+
+    /**
+     * @param input Starts a new workflow execution
+     * @return id of the workflow
+     */
+    String startWorkflow(StartWorkflowInput input);
 }
