+++ core/src/main/java/com/netflix/conductor/core/execution/tasks/StartWorkflow.java
@@ -24,7 +24,6 @@ import com.netflix.conductor.common.metadata.workflow.StartWorkflowRequest;
 import com.netflix.conductor.core.exception.TransientException;
 import com.netflix.conductor.core.execution.StartWorkflowInput;
 import com.netflix.conductor.core.execution.WorkflowExecutor;
-import com.netflix.conductor.core.operation.StartWorkflowOperation;
 import com.netflix.conductor.model.TaskModel;
 import com.netflix.conductor.model.WorkflowModel;
 
@@ -45,16 +44,11 @@ public class StartWorkflow extends WorkflowSystemTask {
 
     private final ObjectMapper objectMapper;
     private final Validator validator;
-    private final StartWorkflowOperation startWorkflowOperation;
 
-    public StartWorkflow(
-            ObjectMapper objectMapper,
-            Validator validator,
-            StartWorkflowOperation startWorkflowOperation) {
+    public StartWorkflow(ObjectMapper objectMapper, Validator validator) {
         super(TASK_TYPE_START_WORKFLOW);
         this.objectMapper = objectMapper;
         this.validator = validator;
-        this.startWorkflowOperation = startWorkflowOperation;
     }
 
     @Override
@@ -78,7 +72,7 @@ public class StartWorkflow extends WorkflowSystemTask {
                         request.getCorrelationId(), workflow.getCorrelationId()));
 
         try {
-            String workflowId = startWorkflow(request, workflow.getWorkflowId());
+            String workflowId = startWorkflow(request, workflow.getWorkflowId(), workflowExecutor);
             taskModel.addOutput(WORKFLOW_ID, workflowId);
             taskModel.setStatus(COMPLETED);
         } catch (TransientException te) {
@@ -143,10 +137,11 @@ public class StartWorkflow extends WorkflowSystemTask {
         return startWorkflowRequest;
     }
 
-    private String startWorkflow(StartWorkflowRequest request, String workflowId) {
+    private String startWorkflow(
+            StartWorkflowRequest request, String workflowId, WorkflowExecutor workflowExecutor) {
         StartWorkflowInput input = new StartWorkflowInput(request);
         input.setTriggeringWorkflowId(workflowId);
-        return startWorkflowOperation.execute(input);
+        return workflowExecutor.startWorkflow(input);
     }
 
     @Override
