--- core/src/main/java/com/netflix/conductor/core/operation/WorkflowOperation.java
@@ -1,18 +1,0 @@
-/*
- * Copyright 2022 Conductor Authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package com.netflix.conductor.core.operation;
-
-public interface WorkflowOperation<T, R> {
-
-    R execute(T input);
-}
