+++ core/src/test/groovy/com/netflix/conductor/core/execution/tasks/StartWorkflowSpec.groovy
@@ -16,7 +16,6 @@ import com.netflix.conductor.common.config.ObjectMapperProvider
 import com.netflix.conductor.core.exception.NotFoundException
 import com.netflix.conductor.core.exception.TransientException
 import com.netflix.conductor.core.execution.WorkflowExecutor
-import com.netflix.conductor.core.operation.StartWorkflowOperation
 import com.netflix.conductor.model.TaskModel
 import com.netflix.conductor.model.WorkflowModel
 
@@ -41,21 +40,19 @@ class StartWorkflowSpec extends Specification {
     Validator validator
     WorkflowModel workflowModel
     TaskModel taskModel
-    StartWorkflowOperation startWorkflowOperation
 
     def setup() {
         workflowExecutor = Mock(WorkflowExecutor.class)
         validator = Mock(Validator.class) {
             validate(_) >> new HashSet<ConstraintViolation<Object>>()
         }
-        startWorkflowOperation = Mock(StartWorkflowOperation.class)
 
         def inputData = [:]
         inputData[START_WORKFLOW_PARAMETER] = ['name': 'some_workflow']
         taskModel = new TaskModel(status: SCHEDULED, inputData: inputData)
         workflowModel = new WorkflowModel()
 
-        startWorkflow = new StartWorkflow(new ObjectMapperProvider().getObjectMapper(), validator, startWorkflowOperation)
+        startWorkflow = new StartWorkflow(new ObjectMapperProvider().getObjectMapper(), validator)
     }
 
     def "StartWorkflow task is asynchronous"() {
@@ -93,7 +90,7 @@ class StartWorkflowSpec extends Specification {
 
         then:
         taskModel.status == SCHEDULED
-        1 * startWorkflowOperation.execute(*_) >> { throw new TransientException("") }
+        1 * workflowExecutor.startWorkflow(*_) >> { throw new TransientException("") }
     }
 
     def "WorkflowExecutor throws a NotFoundException"() {
@@ -103,7 +100,7 @@ class StartWorkflowSpec extends Specification {
         then:
         taskModel.status == FAILED
         taskModel.reasonForIncompletion != null
-        1 * startWorkflowOperation.execute(*_) >> { throw new NotFoundException("") }
+        1 * workflowExecutor.startWorkflow(*_) >> { throw new NotFoundException("") }
     }
 
     def "WorkflowExecutor throws a RuntimeException"() {
@@ -113,6 +110,6 @@ class StartWorkflowSpec extends Specification {
         then:
         taskModel.status == FAILED
         taskModel.reasonForIncompletion != null
-        1 * startWorkflowOperation.execute(*_) >> { throw new RuntimeException("I am an unexpected exception") }
+        1 * workflowExecutor.startWorkflow(*_) >> { throw new RuntimeException("I am an unexpected exception") }
     }
 }
