+++ core/src/test/java/com/netflix/conductor/core/events/TestDefaultEventProcessor.java
@@ -44,7 +44,6 @@ import com.netflix.conductor.core.execution.StartWorkflowInput;
 import com.netflix.conductor.core.execution.WorkflowExecutor;
 import com.netflix.conductor.core.execution.evaluators.Evaluator;
 import com.netflix.conductor.core.execution.evaluators.JavascriptEvaluator;
-import com.netflix.conductor.core.operation.StartWorkflowOperation;
 import com.netflix.conductor.core.utils.ExternalPayloadStorageUtils;
 import com.netflix.conductor.core.utils.JsonUtils;
 import com.netflix.conductor.core.utils.ParametersUtils;
@@ -73,7 +72,6 @@ public class TestDefaultEventProcessor {
     private MetadataService metadataService;
     private ExecutionService executionService;
     private WorkflowExecutor workflowExecutor;
-    private StartWorkflowOperation startWorkflowOperation;
     private ExternalPayloadStorageUtils externalPayloadStorageUtils;
     private SimpleActionProcessor actionProcessor;
     private ParametersUtils parametersUtils;
@@ -100,7 +98,6 @@ public class TestDefaultEventProcessor {
         metadataService = mock(MetadataService.class);
         executionService = mock(ExecutionService.class);
         workflowExecutor = mock(WorkflowExecutor.class);
-        startWorkflowOperation = mock(StartWorkflowOperation.class);
         externalPayloadStorageUtils = mock(ExternalPayloadStorageUtils.class);
         actionProcessor = mock(SimpleActionProcessor.class);
         parametersUtils = new ParametersUtils(objectMapper);
@@ -170,8 +167,8 @@ public class TestDefaultEventProcessor {
                                     started.set(true);
                                     return id;
                                 })
-                .when(startWorkflowOperation)
-                .execute(
+                .when(workflowExecutor)
+                .startWorkflow(
                         argThat(
                                 argument ->
                                         startWorkflowAction
@@ -204,8 +201,7 @@ public class TestDefaultEventProcessor {
         doNothing().when(externalPayloadStorageUtils).verifyAndUpload(any(), any());
 
         SimpleActionProcessor actionProcessor =
-                new SimpleActionProcessor(
-                        workflowExecutor, parametersUtils, jsonUtils, startWorkflowOperation);
+                new SimpleActionProcessor(workflowExecutor, parametersUtils, jsonUtils);
 
         DefaultEventProcessor eventProcessor =
                 new DefaultEventProcessor(
@@ -262,8 +258,8 @@ public class TestDefaultEventProcessor {
                                     started.set(true);
                                     return id;
                                 })
-                .when(startWorkflowOperation)
-                .execute(
+                .when(workflowExecutor)
+                .startWorkflow(
                         argThat(
                                 argument ->
                                         startWorkflowAction
@@ -277,8 +273,7 @@ public class TestDefaultEventProcessor {
                                                 && event.equals(argument.getEvent())));
 
         SimpleActionProcessor actionProcessor =
-                new SimpleActionProcessor(
-                        workflowExecutor, parametersUtils, jsonUtils, startWorkflowOperation);
+                new SimpleActionProcessor(workflowExecutor, parametersUtils, jsonUtils);
 
         DefaultEventProcessor eventProcessor =
                 new DefaultEventProcessor(
@@ -332,8 +327,8 @@ public class TestDefaultEventProcessor {
                                     started.set(true);
                                     return id;
                                 })
-                .when(startWorkflowOperation)
-                .execute(
+                .when(workflowExecutor)
+                .startWorkflow(
                         argThat(
                                 argument ->
                                         startWorkflowAction
@@ -347,8 +342,7 @@ public class TestDefaultEventProcessor {
                                                 && event.equals(argument.getEvent())));
 
         SimpleActionProcessor actionProcessor =
-                new SimpleActionProcessor(
-                        workflowExecutor, parametersUtils, jsonUtils, startWorkflowOperation);
+                new SimpleActionProcessor(workflowExecutor, parametersUtils, jsonUtils);
 
         DefaultEventProcessor eventProcessor =
                 new DefaultEventProcessor(
