+++ core/src/test/java/com/netflix/conductor/core/events/TestSimpleActionProcessor.java
@@ -33,7 +33,6 @@ import com.netflix.conductor.common.metadata.tasks.TaskResult.Status;
 import com.netflix.conductor.common.metadata.workflow.WorkflowDef;
 import com.netflix.conductor.core.execution.StartWorkflowInput;
 import com.netflix.conductor.core.execution.WorkflowExecutor;
-import com.netflix.conductor.core.operation.StartWorkflowOperation;
 import com.netflix.conductor.core.utils.ExternalPayloadStorageUtils;
 import com.netflix.conductor.core.utils.JsonUtils;
 import com.netflix.conductor.core.utils.ParametersUtils;
@@ -53,7 +52,6 @@ public class TestSimpleActionProcessor {
     private WorkflowExecutor workflowExecutor;
     private ExternalPayloadStorageUtils externalPayloadStorageUtils;
     private SimpleActionProcessor actionProcessor;
-    private StartWorkflowOperation startWorkflowOperation;
 
     @Autowired private ObjectMapper objectMapper;
 
@@ -62,14 +60,12 @@ public class TestSimpleActionProcessor {
         externalPayloadStorageUtils = mock(ExternalPayloadStorageUtils.class);
 
         workflowExecutor = mock(WorkflowExecutor.class);
-        startWorkflowOperation = mock(StartWorkflowOperation.class);
 
         actionProcessor =
                 new SimpleActionProcessor(
                         workflowExecutor,
                         new ParametersUtils(objectMapper),
-                        new JsonUtils(objectMapper),
-                        startWorkflowOperation);
+                        new JsonUtils(objectMapper));
     }
 
     @SuppressWarnings({"unchecked", "rawtypes"})
@@ -96,7 +92,7 @@ public class TestSimpleActionProcessor {
         workflowDef.setName("testWorkflow");
         workflowDef.setVersion(1);
 
-        when(startWorkflowOperation.execute(any())).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(any())).thenReturn("workflow_1");
 
         Map<String, Object> output =
                 actionProcessor.execute(action, payload, "testEvent", "testMessage");
@@ -107,7 +103,7 @@ public class TestSimpleActionProcessor {
         ArgumentCaptor<StartWorkflowInput> startWorkflowInputArgumentCaptor =
                 ArgumentCaptor.forClass(StartWorkflowInput.class);
 
-        verify(startWorkflowOperation).execute(startWorkflowInputArgumentCaptor.capture());
+        verify(workflowExecutor).startWorkflow(startWorkflowInputArgumentCaptor.capture());
         StartWorkflowInput capturedValue = startWorkflowInputArgumentCaptor.getValue();
 
         assertEquals("test_1", capturedValue.getWorkflowInput().get("testInput"));
@@ -139,7 +135,7 @@ public class TestSimpleActionProcessor {
         workflowDef.setName("testWorkflow");
         workflowDef.setVersion(1);
 
-        when(startWorkflowOperation.execute(any())).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(any())).thenReturn("workflow_1");
 
         Map<String, Object> output =
                 actionProcessor.execute(action, payload, "testEvent", "testMessage");
@@ -150,7 +146,7 @@ public class TestSimpleActionProcessor {
         ArgumentCaptor<StartWorkflowInput> startWorkflowInputArgumentCaptor =
                 ArgumentCaptor.forClass(StartWorkflowInput.class);
 
-        verify(startWorkflowOperation).execute(startWorkflowInputArgumentCaptor.capture());
+        verify(workflowExecutor).startWorkflow(startWorkflowInputArgumentCaptor.capture());
         StartWorkflowInput capturedArgument = startWorkflowInputArgumentCaptor.getValue();
         assertEquals("test_1", capturedArgument.getWorkflowInput().get("testInput"));
         assertNull(capturedArgument.getCorrelationId());
