+++ core/src/test/java/com/netflix/conductor/core/execution/tasks/TestSubWorkflow.java
@@ -28,7 +28,6 @@ import com.netflix.conductor.core.exception.NonTransientException;
 import com.netflix.conductor.core.exception.TransientException;
 import com.netflix.conductor.core.execution.StartWorkflowInput;
 import com.netflix.conductor.core.execution.WorkflowExecutor;
-import com.netflix.conductor.core.operation.StartWorkflowOperation;
 import com.netflix.conductor.model.TaskModel;
 import com.netflix.conductor.model.WorkflowModel;
 
@@ -49,15 +48,13 @@ public class TestSubWorkflow {
 
     private WorkflowExecutor workflowExecutor;
     private SubWorkflow subWorkflow;
-    private StartWorkflowOperation startWorkflowOperation;
 
     @Autowired private ObjectMapper objectMapper;
 
     @Before
     public void setup() {
         workflowExecutor = mock(WorkflowExecutor.class);
-        startWorkflowOperation = mock(StartWorkflowOperation.class);
-        subWorkflow = new SubWorkflow(objectMapper, startWorkflowOperation);
+        subWorkflow = new SubWorkflow(objectMapper);
     }
 
     @Test
@@ -84,7 +81,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn(workflowId);
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn(workflowId);
 
         when(workflowExecutor.getWorkflow(anyString(), eq(false))).thenReturn(workflow);
 
@@ -125,7 +122,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput))
+        when(workflowExecutor.startWorkflow(startWorkflowInput))
                 .thenThrow(new TransientException("QueueDAO failure"));
 
         subWorkflow.start(workflowInstance, task, workflowExecutor);
@@ -156,7 +153,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
         String failureReason = "non transient failure";
-        when(startWorkflowOperation.execute(startWorkflowInput))
+        when(workflowExecutor.startWorkflow(startWorkflowInput))
                 .thenThrow(new NonTransientException(failureReason));
 
         subWorkflow.start(workflowInstance, task, workflowExecutor);
@@ -189,7 +186,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
 
         subWorkflow.start(workflowInstance, task, workflowExecutor);
         assertEquals("workflow_1", task.getSubWorkflowId());
@@ -219,7 +216,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(workflowInput);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
 
         subWorkflow.start(workflowInstance, task, workflowExecutor);
         assertEquals("workflow_1", task.getSubWorkflowId());
@@ -252,7 +249,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(taskToDomain);
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
 
         subWorkflow.start(workflowInstance, task, workflowExecutor);
         assertEquals("workflow_1", task.getSubWorkflowId());
@@ -278,7 +275,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
 
         assertFalse(subWorkflow.execute(workflowInstance, task, workflowExecutor));
     }
@@ -313,7 +310,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(taskToDomain);
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
         when(workflowExecutor.getWorkflow(eq("sub-workflow-id"), eq(false)))
                 .thenReturn(subWorkflowInstance);
 
@@ -372,7 +369,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
         when(workflowExecutor.getWorkflow(eq("sub-workflow-id"), eq(true)))
                 .thenReturn(subWorkflowInstance);
 
@@ -404,7 +401,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowInput(inputData);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
         when(workflowExecutor.getWorkflow(eq("sub-workflow-id"), eq(false)))
                 .thenReturn(subWorkflowInstance);
 
@@ -443,7 +440,7 @@ public class TestSubWorkflow {
         startWorkflowInput.setWorkflowDefinition(subWorkflowDef);
         startWorkflowInput.setTaskToDomain(workflowInstance.getTaskToDomain());
 
-        when(startWorkflowOperation.execute(startWorkflowInput)).thenReturn("workflow_1");
+        when(workflowExecutor.startWorkflow(startWorkflowInput)).thenReturn("workflow_1");
 
         subWorkflow.start(workflowInstance, task, workflowExecutor);
         assertEquals("workflow_1", task.getSubWorkflowId());
