+++ core/src/test/java/com/netflix/conductor/service/WorkflowServiceTest.java
@@ -34,7 +34,6 @@ import com.netflix.conductor.common.run.Workflow;
 import com.netflix.conductor.common.run.WorkflowSummary;
 import com.netflix.conductor.core.exception.NotFoundException;
 import com.netflix.conductor.core.execution.WorkflowExecutor;
-import com.netflix.conductor.core.operation.StartWorkflowOperation;
 
 import jakarta.validation.ConstraintViolationException;
 
@@ -61,11 +60,6 @@ public class WorkflowServiceTest {
         }
 
         @Bean
-        public StartWorkflowOperation startWorkflowOperation() {
-            return mock(StartWorkflowOperation.class);
-        }
-
-        @Bean
         public ExecutionService executionService() {
             return mock(ExecutionService.class);
         }
@@ -79,10 +73,8 @@ public class WorkflowServiceTest {
         public WorkflowService workflowService(
                 WorkflowExecutor workflowExecutor,
                 ExecutionService executionService,
-                MetadataService metadataService,
-                StartWorkflowOperation startWorkflowOperation) {
-            return new WorkflowServiceImpl(
-                    workflowExecutor, executionService, metadataService, startWorkflowOperation);
+                MetadataService metadataService) {
+            return new WorkflowServiceImpl(workflowExecutor, executionService, metadataService);
         }
     }
