+++ test-harness/src/test/groovy/com/netflix/conductor/test/base/AbstractSpecification.groovy
@@ -20,7 +20,6 @@ import com.netflix.conductor.ConductorTestApp
 import com.netflix.conductor.core.execution.AsyncSystemTaskExecutor
 import com.netflix.conductor.core.execution.StartWorkflowInput
 import com.netflix.conductor.core.execution.WorkflowExecutor
-import com.netflix.conductor.core.operation.StartWorkflowOperation
 import com.netflix.conductor.core.reconciliation.WorkflowSweeper
 import com.netflix.conductor.service.ExecutionService
 import com.netflix.conductor.service.MetadataService
@@ -53,9 +52,6 @@ abstract class AbstractSpecification extends Specification {
     @Autowired
     AsyncSystemTaskExecutor asyncSystemTaskExecutor
 
-    @Autowired
-    StartWorkflowOperation startWorkflowOperation
-
     def cleanup() {
         workflowTestUtil.clearWorkflows()
     }
@@ -67,6 +63,6 @@ abstract class AbstractSpecification extends Specification {
     protected String startWorkflow(String name, Integer version, String correlationId, Map<String, Object> workflowInput, String workflowInputPath) {
         StartWorkflowInput input = new StartWorkflowInput(name: name, version: version, correlationId: correlationId, workflowInput: workflowInput, externalInputPayloadStoragePath: workflowInputPath)
 
-        startWorkflowOperation.execute(input)
+        workflowExecutor.startWorkflow(input)
     }
 }
